backend/requirements.txt
# backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
spacy==3.7.6
pdfplumber==0.11.4
python-docx==1.1.2
openai==1.46.1
python-dotenv==1.0.1
loguru==0.7.2
pytest==8.3.3
httpx==0.27.2
reportlab==4.2.2